app/requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
orjson==3.9.0
packaging==23.1
pandas==2.0.2
pluggy==1.0.0
pydantic==1.10.8
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.6.2
tzdata==2023.3
ujson==5.7.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3